requirements.txt
opencv-python
mediapipe
pyserial